backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiomysql==0.2.0
pymysql==1.1.0
alembic==1.13.1
cryptography==46.0.5

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation and Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client (for AI API)
httpx==0.26.0
aiohttp==3.9.1

# Google Gemini AI
google-generativeai==0.3.2

# File uploads
aiofiles==23.2.1

# Utilities
python-slugify==8.0.1
python-dotenv==1.0.0
markdown==3.5.1
bleach==6.1.0

# Caching
redis==5.0.1

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
